imported def activity and choice instead of having them at the same place
